Flask_App/requirements.txt
pymongo 
flask
jsonify
google-generativeai
gunicorn